requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
sentence-transformers==5.1.2
faiss-cpu==1.13.0
python-dotenv==1.0.1
pymongo==4.15.4
google-generativeai==0.8.5
langchain-text-splitters==0.3.2
pydantic==2.12.4
numpy==2.2.6